refactor(amdsmi-tests): sort functional C++ stragglers into subdirs
Five functional test files were left flat in functional/ after the rebase instead
of being sorted into the refactor's per-area subdirectories. Move each .cc/.h pair
into the matching subdir and fix the now-shorter test_base.h/test_common.h includes
("../test_base.h" -> "test_base.h"; the test root is already on the include path).

- gpu/partition/: computepartition_memallocmode_read_write (with the other partition tests)
- system/:        kfd_atfork_read, cross_process_serialization (process/runtime behavior)
- ifoe/:          fabric_read (UALoE fabric runs over IFoE, not XGMI), ifoe_info_read

Also fix a latent gap: ifoe_info_read defined a TestBase subclass but was never
registered in main.cc, so it never ran. Add its TEST(GpuFunctionalReadOnly,
TestIfoeInfoRead) registration alongside the other ifoe test. Remove the now-redundant
functional/ifoe/placeholder.cc (a real test lives there now).

main.cc include paths updated to match. Both unit/ and functional/ glob into the same
amdsmitst binary, so the moves are purely categorization; rebuilt amdsmitst and verified
all tests register (TestIfoeInfoRead newly appears, 44 -> 45 C++ tests).
