+++ projects/amdsmi/tests/amd_smi_test/functional/gpu/partition/computepartition_memallocmode_read_write.cc
@@ -27,8 +27,8 @@
 #include <string>
 #include <vector>
 
-#include "../test_base.h"
-#include "../test_common.h"
+#include "test_base.h"
+#include "test_common.h"
 #include "amd_smi/amdsmi.h"
 
 TestComputePartitionMemAllocModeReadWrite::TestComputePartitionMemAllocModeReadWrite()
