+++ projects/amdsmi/tests/amd_smi_test/functional/ifoe/fabric_read.cc
