+++ projects/amdsmi/tests/amd_smi_test/functional/ifoe/ifoe_info_read.cc
@@ -30,7 +30,7 @@
 #include <map>
 #include <string>
 
-#include "../test_common.h"
+#include "test_common.h"
 #include "amd_smi/amdsmi.h"
 
 TestIfoeInfoRead::TestIfoeInfoRead() : TestBase() {
