+++ projects/amdsmi/tests/amd_smi_test/functional/system/cross_process_serialization.cc
@@ -35,7 +35,7 @@
 #include <string>
 #include <vector>
 
-#include "../test_common.h"
+#include "test_common.h"
 #include "amd_smi/amdsmi.h"
 #include "amd_smi/impl/amd_smi_test_internal.h"
 #include "rocm_smi/rocm_smi_utils.h"
